pyproject.toml
[tool.poetry]
name = "falconcv"
version = "1.0.5"
description = ""
authors = ["Henry Ruiz <henry.ruiz@tamu.edu>"]

[tool.poetry.dependencies]
#python = ">=3.7,<=3.8.*"
python = "3.7.*"
dask = {version= "2.30.0", extras = ["complete"]}
rich = "^12.4.4"
#tensorflow = "^2.9.1"
requests = "^2.28.0"
Mako = "^1.2.0"
boto3 = "^1.24.13"
tflite-support = "^0.4.1"
bs4 = "^0.0.1"
validators = "^0.20.0"
GitPython = "^3.1.27"
alive-progress = "^2.4.1"
Pillow = "^9.1.1"
matplotlib = "^3.5.2"
pandas="<=1.3.5"
seaborn = "^0.11.2"
lxml = "^4.9.0"
typing-extensions = "^4.2.0"
opencv-python = "^4.6.0"
pyparsing = "2.4.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
python-dotenv = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
